fms-gateway/src/main/webapp/app/entities/flights/flight/flight.service.ts
import { Injectable } from '@angular/core';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';

import { SERVER_API_URL } from 'app/app.constants';
import { createRequestOption } from 'app/shared/util/request-util';
import { IFlight } from 'app/shared/model/flights/flight.model';

type EntityResponseType = HttpResponse<IFlight>;
type EntityArrayResponseType = HttpResponse<IFlight[]>;

@Injectable({ providedIn: 'root' })
export class FlightService {
  public resourceUrl = SERVER_API_URL + 'services/flights/api/flights';

  constructor(protected http: HttpClient) {}

  create(flight: IFlight): Observable<EntityResponseType> {
    return this.http.post<IFlight>(this.resourceUrl, flight, { observe: 'response' });
  }

  update(flight: IFlight): Observable<EntityResponseType> {
    return this.http.put<IFlight>(this.resourceUrl, flight, { observe: 'response' });
  }

  find(id: number): Observable<EntityResponseType> {
    return this.http.get<IFlight>(`${this.resourceUrl}/${id}`, { observe: 'response' });
  }

  query(req?: any): Observable<EntityArrayResponseType> {
    const options = createRequestOption(req);
    return this.http.get<IFlight[]>(this.resourceUrl, { params: options, observe: 'response' });
  }

  delete(id: number): Observable<HttpResponse<{}>> {
    return this.http.delete(`${this.resourceUrl}/${id}`, { observe: 'response' });
  }
}
